# docs/swagger.yaml
basePath: /v1
definitions:
...
  dto.ForgotPasswordRequest:
    properties:
      phone_number:
        description: PhoneNumber is the Ethiopian phone number of the user
        type: string
    type: object
  dto.ForgotPasswordResponse:
    properties:
      message:
        description: Message is a generic success message (doesn't reveal if user
          exists)
        type: string
    type: object
...
  dto.ResetPasswordRequest:
    properties:
      new_password:
        description: NewPassword is the new password to set
        type: string
      token:
        description: Token is the password reset token
        type: string
    type: object
  dto.ResetPasswordResponse:
    properties:
      message:
        description: Message is the success message
        type: string
    type: object
  dto.TelegramLinkRequest:
    properties:
      phone_number:
        description: PhoneNumber is the user's phone number
        type: string
    type: object
  dto.TelegramLinkResponse:
    properties:
      link_code:
        description: LinkCode is the one-time code for linking Telegram account
        type: string
      message:
        description: Message provides instructions to the user
        type: string
      telegram_url:
        description: TelegramURL is the deep link URL to open in Telegram
        type: string
    type: object
...
  model.ErrorResponse:
    properties:
      code:
        description: Code is the error code. It is not status code
        type: integer
      description:
        description: Description is the error description.
        type: string
      field_error:
        description: FieldError is the error detail for each field, if available that
          is.
        items:
          $ref: '#/definitions/model.FieldError'
        type: array
      message:
        description: Message is the error message.
        type: string
      stack_trace:
        description: |-
          StackTrace is the stack trace of the error.
          It is only returned for debugging
        type: string
    type: object
  model.FieldError:
    properties:
      description:
        description: Description is the error description for this field.
        type: string
      name:
        description: Name is the name of the field that caused the error.
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: A comprehensive authentication API supporting Ethiopian phone numbers
    with JWT-based security
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Kesis Ethiopian Phone Authentication API
  version: "1.0"
paths:
...
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Initiate password reset process. A reset link will be sent to the
        user's linked Telegram account if it exists
      parameters:
      - description: Forgot password request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success message (always returned regardless of user existence)
          schema:
            $ref: '#/definitions/dto.ForgotPasswordResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Request password reset
      tags:
      - Authentication
  /auth/link-telegram:
    post:
      consumes:
      - application/json
      description: Generate a temporary link code to connect user's Telegram account
      parameters:
      - description: Telegram link request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.TelegramLinkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Link code and Telegram URL generated
          schema:
            $ref: '#/definitions/dto.TelegramLinkResponse'
        "400":
          description: User not found or validation error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Generate Telegram link code
      tags:
      - Authentication
...
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset user password using a valid reset token
      parameters:
      - description: Reset password request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset successful
          schema:
            $ref: '#/definitions/dto.ResetPasswordResponse'
        "400":
          description: Invalid token or validation error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Complete password reset
      tags:
      - Authentication
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter the token with the `Bearer: ` prefix, e.g. "Bearer abcde12345"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"